Fix error when tokenizing unicode chars
As per https://github.com/pndurette/gTTS/issues/21 non ASCII chars were not correctly being calculated. This due to an error in the calculate_token function. The diff should be self explanatory when comparing it with the token-script.js. I added a couple of tests to make sure all unicode chars are correctly being calculated.