Remove client-side check for language support
Don't check for language support on the client side, since the list
of supported languages is rapidly evolving, and even some undocumented
language codes are accepted, e.g. "de" for German, in addition to the
documented "de-DE".

By not checking the language code on the client side, using an invalid
code will raise a 404 exception when the request is made:

    requests.exceptions.HTTPError: 404 Client Error: <snip>